+++ outputs.tf
@@ -1,12 +1,16 @@
 output "bucket_name" {
   value = local.bucket_name
 }
+output "bucket_path" {
+  value = var.bucket_path
+}
 output "bucket_queue_name" {
   value = local.bucket_queue_name
 }
 output "bucket_region" {
   value = data.aws_s3_bucket.file_storage.region
 }
+
 output "cluster_id" {
   value = module.app_eks.cluster_id
 }